# ressources/openAPI.yml
openapi: 3.0.3
info:
  title: EventSync API
  description: >
    API REST de la plateforme EventSync — gestion d'événements et engagement
    des participants en temps réel. Remplace les supports statiques par une
    interface dynamique avec planning multi-track, sessions live, Q&R et favoris.
  version: 1.0.0
  contact:
    name: EventSync Team
servers:
  - url: http://localhost:3000
    description: Serveur de production (backend)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT obtenu via /auth/login. 
        Utilisé uniquement pour les routes admin (organisateur).

  # -------------------------------------------------------------------------
  # SCHEMAS
  # -------------------------------------------------------------------------
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: admin@gmail.com
        password:
          type: string
          format: password
          example: "password"
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT token à utiliser dans Authorization header
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "Bearer"
        expires_in:
          type: string
          example: "1d"

    # ── Event ───────────────────────────────────────────────────────────────
...
    Error:
      type: object
      properties:
        code:
          type: string
          example: "NOT_FOUND"
        message:
          type: string
          example: "La ressource demandée est introuvable."
...
paths:
...
  /auth/login:
    post:
      tags: [ Auth ]
      summary: Connexion organisateur (admin)
      description: >
        Authentifie un organisateur et retourne un JWT pour accéder aux routes protégées.
      operationId: loginOrganizer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Identifiants invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  # ══════════════════════════════════════════════════════════════════════════
  # EVENTS
  # ══════════════════════════════════════════════════════════════════════════
...